DA-24: - Refactored testcontainers configuration - Refactored entity classes

+++ src/main/kotlin/com/dataart/coreservice/repository/EventRepository.kt
@@ -3,4 +3,4 @@ package com.dataart.coreservice.repository
 import com.dataart.coreservice.model.Event
 import org.springframework.data.repository.CrudRepository
 
-interface EventRepository : CrudRepository<Event, Int>
+interface EventRepository : CrudRepository<Event, Long>
